requirements.txt
vk_api
python-dotenv
detoxify
torch
sentencepiece
transformers